fix: add handler for checkbox and radio button state change
