+++ Sources/Public/Components/SelectionControl/NatSelectorProtocol.swift
@@ -1,0 +1,10 @@
+public typealias SelectionHandler = ((Bool) -> Void)?
+
+protocol NatSelector: UIControl, Pulsable {
+    var isHapticFeedbackEnabled: Bool { get set }
+    var isIndeterminate: Bool { get set }
+    var labelComponent: String? { get set }
+    var isGrouped: Bool { get set }
+    var groupId: Int { get set }
+    var handler: SelectionHandler { get set }
+}
