+++ Sources/Public/Components/SelectionControl/Selectors/NatCheckboxContainer.swift
@@ -1,13 +1,3 @@
-import UIKit
-
-protocol NatSelector: UIControl, Pulsable {
-    var isHapticFeedbackEnabled: Bool { get set }
-    var isIndeterminate: Bool { get set }
-    var labelComponent: String? { get set }
-    var isGrouped: Bool { get set }
-    var groupId: Int { get set }
-}
-
 class NatCheckboxContainer: UIControl, NatSelector {
 
     let checkbox = NatCheckboxControl()
@@ -29,6 +19,15 @@ class NatCheckboxContainer: UIControl, NatSelector {
         }
         set {
             checkbox.isEnabled = newValue
+        }
+    }
+
+    var handler: SelectionHandler {
+        get {
+            checkbox.handler
+        }
+        set {
+            checkbox.handler = newValue
         }
     }
