+++ Sources/Public/Components/SelectionControl/Selectors/NatRadioButton/NatRadioButtonControl.swift
@@ -5,6 +5,7 @@ final class NatRadioButtonControl: UIControl {
     var onTouchesBegan: ((Set<UITouch>) -> Void)?
     var onTouchesEnded: ((Set<UITouch>) -> Void)?
 
+    var handler: ((Bool) -> Void)?
     var isHapticFeedbackEnabled: Bool = false
     var isIndeterminate: Bool = false
     var labelComponent: String?
@@ -20,6 +21,7 @@ final class NatRadioButtonControl: UIControl {
         didSet {
             self.setNeedsDisplay() 
             super.isSelected = isSelected
+            handler?(isSelected)
         }
     }
